Refactor LargeHeroSection into own component file
